Removed 3 code smell
Renamed variables, split methods.

HttpHeaderFilterRegexpTest added test that checks that remove regex not removes "Transfer_Encoding" This is sort of odd test to have, but because it makes it clear that failing test mentioned in (NTP-925) that is related to new VP not returning that header I ceep that test for now (It is at least not harmful)
WsdlConfigurationJson introduce a F_SLASH constant instead of "/" to remove code smell. (note that what is done here is NOT making a path to file system so having forward slash is totally Ok in all instances)

RivTaProfilProcessor added a interface and anonymous lambda "to save a few parameters" that Sonarqube thought was one to many. Not sure if this was best solution but fixed the complaint.